Arel::Nodes::Similar
