+++ lib/arel/extensions.rb
@@ -29,6 +29,7 @@ require 'arel/extensions/type_cast'
 require 'arel/extensions/distinct_from'
 require 'arel/extensions/not_distinct_from'
 require 'arel/extensions/null_if'
+require 'arel/extensions/similar'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
